requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
discord.py==1.7.3
idna==3.2
multidict==5.1.0
random-cat==1.0.1
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
AnilistPython==0.1.3
letterboxdpy==2.1
numpy==1.23.2
lxml==4.9.1

